+++ crates/voicers-nn/src/albert.rs
@@ -5,8 +5,41 @@ use mlx_rs::module::Module;
 use mlx_rs::nn::{Dropout, DropoutBuilder, Embedding, Gelu, LayerNorm, LayerNormBuilder, Linear, LinearBuilder};
 use mlx_rs::ops::{expand_dims, softmax_axis, zeros_like};
 use mlx_rs::Array;
+use serde::Deserialize;
 
-use crate::config::AlbertConfig;
+#[derive(Debug, Clone, Deserialize)]
+pub struct AlbertConfig {
+    pub num_hidden_layers: i32,
+    pub num_attention_heads: i32,
+    pub hidden_size: i32,
+    pub intermediate_size: i32,
+    pub max_position_embeddings: i32,
+    #[serde(default = "default_embedding_size")]
+    pub embedding_size: i32,
+    #[serde(default = "default_inner_group_num")]
+    pub inner_group_num: i32,
+    #[serde(default = "default_num_hidden_groups")]
+    pub num_hidden_groups: i32,
+    #[serde(default = "default_hidden_dropout_prob")]
+    pub hidden_dropout_prob: f32,
+    #[serde(default = "default_attention_probs_dropout_prob")]
+    pub attention_probs_dropout_prob: f32,
+    #[serde(default = "default_type_vocab_size")]
+    pub type_vocab_size: i32,
+    #[serde(default = "default_layer_norm_eps")]
+    pub layer_norm_eps: f32,
+    #[serde(default = "default_vocab_size")]
+    pub vocab_size: i32,
+}
+
+fn default_embedding_size() -> i32 { 128 }
+fn default_inner_group_num() -> i32 { 1 }
+fn default_num_hidden_groups() -> i32 { 1 }
+fn default_hidden_dropout_prob() -> f32 { 0.1 }
+fn default_attention_probs_dropout_prob() -> f32 { 0.1 }
+fn default_type_vocab_size() -> i32 { 2 }
+fn default_layer_norm_eps() -> f32 { 1e-12 }
+fn default_vocab_size() -> i32 { 30522 }
 
 // ---------------------------------------------------------------------------
 // Input types
