+++ crates/voicers-nn/src/lib.rs
@@ -1,7 +1,7 @@
-pub mod lstm;
-pub mod conv_weighted;
 pub mod ada_norm;
 pub mod albert;
-pub mod text_encoder;
+pub mod conv_weighted;
+pub mod lstm;
 pub mod prosody;
+pub mod text_encoder;
 pub mod vocoder;
