+++ crates/voicers-nn/src/lstm.rs
