+++ crates/voicers/src/config.rs
@@ -1,6 +1,8 @@
 use serde::Deserialize;
 use std::collections::HashMap;
 
+pub use voicers_nn::albert::AlbertConfig;
+
 #[derive(Debug, Clone, Deserialize)]
 pub struct ModelConfig {
     pub istftnet: ISTFTNetConfig,
@@ -23,56 +25,6 @@ pub struct ModelConfig {
 
 fn default_sample_rate() -> i32 {
     24000
-}
-
-#[derive(Debug, Clone, Deserialize)]
-pub struct AlbertConfig {
-    pub num_hidden_layers: i32,
-    pub num_attention_heads: i32,
-    pub hidden_size: i32,
-    pub intermediate_size: i32,
-    pub max_position_embeddings: i32,
-    #[serde(default = "default_embedding_size")]
-    pub embedding_size: i32,
-    #[serde(default = "default_inner_group_num")]
-    pub inner_group_num: i32,
-    #[serde(default = "default_num_hidden_groups")]
-    pub num_hidden_groups: i32,
-    #[serde(default = "default_hidden_dropout_prob")]
-    pub hidden_dropout_prob: f32,
-    #[serde(default = "default_attention_probs_dropout_prob")]
-    pub attention_probs_dropout_prob: f32,
-    #[serde(default = "default_type_vocab_size")]
-    pub type_vocab_size: i32,
-    #[serde(default = "default_layer_norm_eps")]
-    pub layer_norm_eps: f32,
-    #[serde(default = "default_vocab_size")]
-    pub vocab_size: i32,
-}
-
-fn default_embedding_size() -> i32 {
-    128
-}
-fn default_inner_group_num() -> i32 {
-    1
-}
-fn default_num_hidden_groups() -> i32 {
-    1
-}
-fn default_hidden_dropout_prob() -> f32 {
-    0.1
-}
-fn default_attention_probs_dropout_prob() -> f32 {
-    0.1
-}
-fn default_type_vocab_size() -> i32 {
-    2
-}
-fn default_layer_norm_eps() -> f32 {
-    1e-12
-}
-fn default_vocab_size() -> i32 {
-    30522
 }
 
 #[derive(Debug, Clone, Deserialize)]
